iata-corrector-api/requirements.txt
openai>=1.0.0
flask
python-dotenv
